moved cards reusable component into notes dir
