+++ app/javascript/src/components/Dashboard/Notes/Card.jsx
